app/viber/requirements.txt
aiofiles==23.2.1
aiohttp==3.9.1
aiosignal==1.3.1
aiosqlite==0.19.0
attrs==23.2.0
black==23.12.1
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
environs==10.3.0
frozenlist==1.4.1
idna==3.6
Jinja2==3.1.3
MarkupSafe==2.1.4
marshmallow==3.20.2
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.1.0
python-dateutil==2.8.2
python-dotenv==1.0.1
redis==5.0.1
requests==2.31.0
six==1.16.0
urllib3==2.1.0
validators==0.22.0
yarl==1.9.4
